[Feature] Simplify and group existing task in gulpfile Fixes #1820
